refactor(build): Rename project and modules
This commit introduces an update to the project and module names.

The key changes include:

- **Project and Module Renaming:**
  - The root project has been renamed from `a2a-sdk-jvm` to `a2a-jvm` for brevity.
  - Modules have been renamed for consistency and clarity (e.g., `a2a-core` is now `a2a-jvm-core`).
  - Package names have been updated accordingly.
