--- a2a-core/src/main/java/io/github/vchhabra/a2a_jvm/core/Action.java
@@ -1,18 +1,0 @@
-package io.github.vchhabra.a2a_jvm.core;
-
-import java.util.Map;
-
-/**
- * Describes a single action (tool) that an agent can execute.
- *
- * @param name The unique name of the action (e.g., "getWeatherForCity").
- * @param description A human-readable description of what the action does.
- * @param inputSchema A JSON Schema object describing the required input parameters.
- * @param operation A map defining how to invoke this action (e.g., via JSON-RPC).
- */
-public record Action(
-        String name,
-        String description,
-        Map<String, Object> inputSchema,
-        Map<String, Object> operation
-) {}
