--- a2a-examples/src/main/java/io/github/vchhabra/a2a_jvm/a2a_examples/ExampleApplication.java
@@ -1,8 +1,0 @@
-package io.github.vchhabra.a2a_jvm.a2a_examples;
-
-public class ExampleApplication {
-
-    public static void main(String[] args) {
-        System.out.println("Hello World!");
-    }
-}
