+++ a2a-jvm-core/build.gradle.kts
