+++ a2a-jvm-examples/src/main/java/io/github/vchhabra/a2a/jvm/examples/ExampleApplication.java
@@ -1,0 +1,8 @@
+package io.github.vchhabra.a2a.jvm.examples;
+
+public class ExampleApplication {
+
+    public static void main(String[] args) {
+        System.out.println("Hello World!");
+    }
+}
