+++ a2a-jvm-server/build.gradle.kts
@@ -1,0 +1,11 @@
+import plugins.conventions.BuildConstants
+
+plugins {
+    // This module is part of the Spring ecosystem.
+    alias(libs.plugins.jvm.spring.library.convention)
+}
+
+dependencies {
+    implementation(project(BuildConstants.Modules.A2A_JVM_CORE))
+    implementation(libs.spring.boot.starter.web)
+}
