+++ a2a-jvm-transport-jsonrpc/build.gradle.kts
@@ -1,0 +1,7 @@
+plugins {
+    alias(libs.plugins.jvm.core.library.convention)
+}
+
+dependencies {
+
+}
