--- a2a-spring-boot-starter/build.gradle.kts
@@ -1,12 +1,0 @@
-import plugins.conventions.BuildConstants
-
-plugins {
-    // This module is part of the Spring ecosystem.
-    alias(libs.plugins.jvm.spring.library.convention)
-}
-
-dependencies {
-    api(project(BuildConstants.Modules.A2A_CLIENT))
-    api(project(BuildConstants.Modules.A2A_SERVER))
-    api(libs.spring.boot.autoconfigure)
-}
